+++ resources/views/admin/system/index.php
@@ -11,6 +11,7 @@ use Hirtz\Skeleton\Modules\Admin\Controllers\SystemController;
 use Hirtz\Skeleton\Modules\Admin\Widgets\MigrationAlert;
 use Hirtz\Skeleton\Modules\Admin\Widgets\Navs\SystemSubmenu;
 use Hirtz\Skeleton\Modules\Admin\Widgets\Panels\ApplicationInfo;
+use Hirtz\Skeleton\Modules\Admin\Widgets\Panels\ServerInfo;
 use Hirtz\Skeleton\Web\View;
 use Hirtz\Skeleton\Widgets\Navs\Header;
 
@@ -22,3 +23,4 @@ echo Header::make()
 echo SystemSubmenu::make();
 echo MigrationAlert::make();
 echo ApplicationInfo::make();
+echo ServerInfo::make();